chore: rename to `mod.ts`
